custom_yolo_lib.experiments: add model factory
